requirements.txt
flask
openai
flask-cors
python-dotenv
google-generativeai
